requirements.txt
click==6.6
Flask==0.11.1
gevent==1.1.2
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyserial==3.1.1
Werkzeug==0.11.11
